// src/tcgcommerce/modules/product/card/price/dto/product.card.price.dto.ts
export class ProductCardPriceDTO {
    productCardItemId: string;
    productCardPrintingName: string;
    productCardPriceCurrentLowPrice: number;
    productCardPriceCurrentMidPrice: number;
    productCardPriceCurrentHighPrice: number;
    productCardPriceCurrentMarketPrice: number;
    productCardPriceCurrentDirectLowPrice: number;
}